关于标签栏的一系列增强

+++ src/GreenChrome.cpp
@@ -6,6 +6,13 @@
 
 #define MAX_SIZE 32767
 
+bool DoubleClickCloseTab = true;
+bool RightClickCloseTab = true;
+bool KeepLeftmostTab = true;
+bool FastTabSwitch = true;
+
+HMODULE hInstance;
+
 //自动添加引号
 void AppendPath(wchar_t *path, const wchar_t* append)
 {
@@ -17,6 +24,164 @@ void AppendPath(wchar_t *path, const wchar_t* append)
 	wcscat(path, temp);
 
 	free(temp);
+}
+
+HHOOK mouse_hook;
+LRESULT CALLBACK MouseProc(int nCode, WPARAM wParam, LPARAM lParam)
+{
+    static bool close_tab_ing = false;
+    bool close_tab = false;
+    bool keep_tab = false;
+    if (nCode==HC_ACTION)
+    {
+        PMOUSEHOOKSTRUCT pmouse = (PMOUSEHOOKSTRUCT) lParam;
+        POINT pt;
+        pt.x = pmouse->pt.x;
+        pt.y = pmouse->pt.y;
+        ScreenToClient(pmouse->hwnd, &pt);
+
+        int close_region[] = {45, 25};
+        if(wParam==WM_LBUTTONDBLCLK&&DoubleClickCloseTab)
+        {
+            if(pt.y<close_region[IsZoomed(pmouse->hwnd)])
+            {
+                close_tab = true;
+            }
+        }
+
+        if(wParam==WM_RBUTTONUP&&RightClickCloseTab)
+        {
+            int hittest = (int)SendMessage(pmouse->hwnd, WM_NCHITTEST, 0, MAKELONG(pmouse->pt.x, pmouse->pt.y));
+
+            if(hittest==HTCLIENT && pt.y<close_region[IsZoomed(pmouse->hwnd)])
+            {
+                close_tab = true;
+            }
+        }
+
+        if(close_tab && KeepLeftmostTab)
+        {
+            int keep_region[] = {204, 212};
+            if(pt.x<keep_region[IsZoomed(pmouse->hwnd)])
+            {
+                keep_tab = true;
+            }
+        }
+
+        if(wParam==WM_MBUTTONUP)
+        {
+            if(close_tab_ing)
+            {
+                close_tab_ing = false;
+            }
+            else
+            {
+                if(KeepLeftmostTab)
+                {
+                    int keep_region[] = {204, 212};
+                    if(pt.x<keep_region[IsZoomed(pmouse->hwnd)])
+                    {
+                        keep_tab = true;
+                        close_tab = true;
+                    }
+                }
+            }
+        }
+
+        if(wParam==WM_MOUSEWHEEL)
+        {
+            typedef struct tagMOUSEHOOKSTRUCTEX {
+              MOUSEHOOKSTRUCT x;
+              DWORD           mouseData;
+            } MOUSEHOOKSTRUCTEX, *PMOUSEHOOKSTRUCTEX, *LPMOUSEHOOKSTRUCTEX;
+            PMOUSEHOOKSTRUCTEX pwheel = (PMOUSEHOOKSTRUCTEX) lParam;
+
+            if(pt.y<close_region[IsZoomed(pmouse->hwnd)])
+            {
+                int zDelta = GET_WHEEL_DELTA_WPARAM(pwheel->mouseData);
+                if(zDelta>0)
+                {
+                    INPUT input[2];
+                    memset(input, 0, sizeof(input));
+
+                    input[0].type = INPUT_KEYBOARD;
+                    input[1].type = INPUT_KEYBOARD;
+
+                    input[0].ki.wVk = VK_CONTROL;
+                    input[1].ki.wVk = VK_PRIOR;
+
+                    input[0].ki.dwFlags = KEYEVENTF_EXTENDEDKEY;
+                    input[1].ki.dwFlags = KEYEVENTF_EXTENDEDKEY;
+                    SendInput(2, input, sizeof(INPUT));
+
+                    input[0].ki.dwFlags |= KEYEVENTF_KEYUP;
+                    input[1].ki.dwFlags |= KEYEVENTF_KEYUP;
+                    SendInput(2, input, sizeof(INPUT));
+                }
+                else
+                {
+                    INPUT input[2];
+                    memset(input, 0, sizeof(input));
+
+                    input[0].type = INPUT_KEYBOARD;
+                    input[1].type = INPUT_KEYBOARD;
+
+                    input[0].ki.wVk = VK_CONTROL;
+                    input[1].ki.wVk = VK_NEXT;
+
+                    input[0].ki.dwFlags = KEYEVENTF_EXTENDEDKEY;
+                    input[1].ki.dwFlags = KEYEVENTF_EXTENDEDKEY;
+                    SendInput(2, input, sizeof(INPUT));
+
+                    input[0].ki.dwFlags |= KEYEVENTF_KEYUP;
+                    input[1].ki.dwFlags |= KEYEVENTF_KEYUP;
+                    SendInput(2, input, sizeof(INPUT));
+
+                }
+                OutputDebugStringA("WM_MOUSEWHEEL");
+                return 1;
+            }
+        }
+    }
+
+    if(keep_tab)
+    {
+        INPUT input[2];
+        memset(input, 0, sizeof(input));
+
+        input[0].type = INPUT_KEYBOARD;
+        input[1].type = INPUT_KEYBOARD;
+
+        input[0].ki.wVk = VK_CONTROL;
+        input[1].ki.wVk = 'T';
+
+        input[0].ki.dwFlags = KEYEVENTF_EXTENDEDKEY;
+        input[1].ki.dwFlags = KEYEVENTF_EXTENDEDKEY;
+        SendInput(2, input, sizeof(INPUT));
+
+        input[0].ki.dwFlags |= KEYEVENTF_KEYUP;
+        input[1].ki.dwFlags |= KEYEVENTF_KEYUP;
+        SendInput(2, input, sizeof(INPUT));
+    }
+
+    if(close_tab)
+    {
+        INPUT input[1];
+        memset(input, 0, sizeof(input));
+
+        input[0].type = INPUT_MOUSE;
+
+        input[0].mi.dwFlags = MOUSEEVENTF_MIDDLEDOWN;
+        SendInput(1, input, sizeof(INPUT));
+
+        input[0].mi.dwFlags = MOUSEEVENTF_MIDDLEUP;
+        SendInput(1, input, sizeof(INPUT));
+
+        close_tab_ing = true;
+        return 1;
+    }
+
+    return CallNextHookEx(mouse_hook, nCode, wParam, lParam );
 }
 
 //构造新命令行
@@ -92,6 +257,8 @@ void NewCommand(const wchar_t *iniPath,const wchar_t *exePath,const wchar_t *ful
 
     if(first_dll)
     {
+        wchar_t *temp = (wchar_t *)malloc(MAX_SIZE);
+
         wchar_t *line = StartProgram;
         while (line && *line)
         {
@@ -100,17 +267,22 @@ void NewCommand(const wchar_t *iniPath,const wchar_t *exePath,const wchar_t *ful
             PROCESS_INFORMATION pi_ = {0};
             si_.cb = sizeof(STARTUPINFO);
 
-            if (CreateProcessW(NULL, line, NULL, NULL, false, CREATE_NEW_CONSOLE | CREATE_UNICODE_ENVIRONMENT | CREATE_DEFAULT_ERROR_MODE, NULL, 0, &si_, &pi_))
+            ExpandEnvironmentStringsW(line, temp, MAX_SIZE);
+            wchar_t *_temp = replace(temp, L"%app%", exePath);
+
+            if (CreateProcessW(NULL, _temp, NULL, NULL, false, CREATE_NEW_CONSOLE | CREATE_UNICODE_ENVIRONMENT | CREATE_DEFAULT_ERROR_MODE, NULL, 0, &si_, &pi_))
             {
                 programs[programs_count] = pi_.hProcess;
                 programs_count++;
                 if(programs_count>=100) break;
             }
 
+            free(_temp);
+
             line += wcslen(line) + 1;
         }
+        free(temp);
     }
-
 
 	//启动进程
 	STARTUPINFOW si = {0};
@@ -127,9 +299,12 @@ void NewCommand(const wchar_t *iniPath,const wchar_t *exePath,const wchar_t *ful
             WaitForSingleObject(pi.hProcess, INFINITE);
 
             //结束附加启动程序
-            for(int i=0;i<programs_count;i++)
+            if(GetPrivateProfileInt(L"其它设置", L"自动关闭", 1, iniPath)==1)
             {
-                TerminateProcess(programs[i], 0);
+                for(int i=0;i<programs_count;i++)
+                {
+                    TerminateProcess(programs[i], 0);
+                }
             }
 
             //强制杀掉额外程序
@@ -190,6 +365,19 @@ void GreenChrome()
 	//修复Win8上常见的没有注册类错误
 	FixNoRegisteredClass();
 
+    //
+    DoubleClickCloseTab = GetPrivateProfileInt(L"其它设置", L"双击关闭标签", 1, iniPath)==1;
+    RightClickCloseTab = GetPrivateProfileInt(L"其它设置", L"右键关闭标签", 1, iniPath)==1;
+    KeepLeftmostTab = GetPrivateProfileInt(L"其它设置", L"保留最左标签", 1, iniPath)==1;
+    FastTabSwitch = GetPrivateProfileInt(L"其它设置", L"快速标签切换", 1, iniPath)==1;
+    if(DoubleClickCloseTab || RightClickCloseTab || KeepLeftmostTab || FastTabSwitch)
+    {
+        if(!wcsstr(GetCommandLineW(), L"--channel"))
+        {
+            mouse_hook = SetWindowsHookEx(WH_MOUSE, MouseProc, hInstance, GetCurrentThreadId());
+        }
+    }
+
 	//父进程不是Chrome，则需要启动GreenChrome功能
 	wchar_t parentPath[MAX_PATH];
 	if(GetParentPath(parentPath) && _wcsicmp(parentPath, fullPath)!=0)
@@ -203,6 +391,8 @@ EXTERNC BOOL WINAPI DllMain(HINSTANCE hModule, DWORD dwReason, LPVOID pv)
 {
 	if (dwReason == DLL_PROCESS_ATTACH)
 	{
+	    hInstance = hModule;
+
 		//保持系统dll原有功能
 		LoadSysDll(hModule);
